+++ backend/auth-service/src/main/java/com/team/project/manager/controller/ManagerController.java
@@ -1,7 +1,7 @@
 package com.team.project.manager.controller;
 
 import com.team.project.manager.entity.User;
-import com.team.project.manager.service.ManagerService;
+import com.team.project.manager.service.AuthenticationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -15,7 +15,7 @@ import java.util.List;
 @RequestMapping("/manager")
 public class ManagerController {
 
-    ManagerService service;
+    AuthenticationService service;
 
     @GetMapping("/dashboard")
     @PreAuthorize("hasAuthority('MANAGER')")
@@ -23,13 +23,5 @@ public class ManagerController {
         return ResponseEntity.ok("Welcome to the Manager Dashboard");
     }
 
-    @GetMapping("/{teamId}/get-member")
-    public ResponseEntity<List<User>> getMembers(@PathVariable("teamId") Integer teamId){
-        return ResponseEntity.ok(service.getMembers(teamId));
-    }
 
-    @GetMapping("/{userId}/get-user")
-    public ResponseEntity<User> getMember(@PathVariable("userId") Integer userId){
-        return ResponseEntity.ok(service.getMember(userId));
-    }
 }
