+++ backend/auth-service/src/main/java/com/team/project/manager/service/AuthenticationService.java
@@ -71,7 +71,93 @@ public class AuthenticationService {
                 .build();
     }
 
+    public Map<String,Object> adminDashboard(){
+        List<User> users = userRepository.findAll();
+        Map<String,Object> response = new HashMap<>();
+        response.put("users",users);
+        return response;
+    }
 
+    public void delete(User user) {
+        this.userRepository.delete(user);
+    }
+
+
+    public AuthenticationResponse update(User user, RegisterRequest request) {
+
+        if(userRepository.findByEmail(request.getEmail()).isPresent())
+            throw new IllegalArgumentException("Email already exists");
+
+        user.setFirstName(request.getFirstName());
+        user.setLastName(request.getLastName());
+        user.setEmail(request.getEmail());
+        user.setPassword(passwordEncoder.encode(request.getPassword()));
+        user.setRole(request.getRole());
+
+        if(request.getRole()==null)
+            throw new IllegalArgumentException("Role must be provided");
+
+        userRepository.save(user);
+
+        var jwtToken = jwtService.generateToken(user);
+        return AuthenticationResponse.builder()
+                .token(jwtToken)
+                .build();
+    }
+
+    public User assignTeam(Integer id, Integer teamId){
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        user.setTeamId(teamId);
+        userRepository.save(user);
+        return user;
+    }
+
+    public ResponseEntity<String> deleteTeam(Integer teamId) {
+        if(!userRepository.findByTeamId(teamId).isEmpty()){
+            List<User> userList = userRepository.findAll();
+            for(User user : userList){
+                if(user.getTeamId()!=null && user.getTeamId().equals(teamId)) {
+                    user.setTeamId(null);
+                    userRepository.save(user);
+
+                }
+            }
+            return ResponseEntity.ok("Team deleted successfully");
+        }
+        return ResponseEntity.ok("No users in the team");
+    }
+
+    public User addMembers(Integer id, Integer teamId) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        user.setTeamId(teamId);
+        userRepository.save(user);
+        return user;
+    }
+
+    public User removeMembers(Integer id) {
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+
+        user.setTeamId(null);
+        userRepository.save(user);
+        return user;
+
+    }
+
+    public List<User> getMembers(Integer teamId) {
+        List<User> users = userRepository.findAllByTeamId(teamId);
+
+        return users;
+    }
+
+
+    public User getMember(Integer userId) {
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        return user;
+    }
 }
